Backend/requirements.txt
pymongo==4.7.2
google-generativeai==0.5.4
transformers==4.41.1
torch==2.3.0
numpy==1.26.4
requests==2.25.1
openai==0.27.4
retrying==1.3.4
requests-oauthlib==1.3.0
pdf2image==1.14.0
setuptools==70.1.0
python-dotenv==1.0.1